services/api/requirements.txt
fastapi
uvicorn
asyncpg
httpx
pyjwt
structlog
prometheus-fastapi-instrumentator